# reflect: optimize Value.IsZero for struct types