--- a/src/reflect/benchmark_test.go
+++ b/src/reflect/benchmark_test.go
@@ -126,6 +126,9 @@ func BenchmarkIsZero(b *testing.B) {
 	type Int1024 struct {
 		a [1024]int
 	}
+	type Int512 struct {
+		a1, a2, a3, a4, a5, a6, a7, a8, a9, a10, a11, a12, a13, a14, a15, a16 [16]S
+	}
 	s := struct {
 		ArrayComparable      [4]T
 		ArrayIncomparable    [4]_Complex
@@ -137,6 +140,7 @@ func BenchmarkIsZero(b *testing.B) {
 		Struct4Int           Int4
 		ArrayStruct4Int_1024 [256]Int4
 		ArrayChanInt_1024    [1024]chan int
+		StructInt_512        Int512
 	}{}
 	s.ArrayInt_1024_NoZero[512] = 1
 	source := ValueOf(s)